보안 관련 코드 첨부, middleware && construct로 바로 해결 할수 있으니 확인이 필요

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+
 //어떤 데이터베이스 쓰는지
 use App\Models\Category;
 
@@ -11,38 +12,61 @@ class CategoryController extends Controller
     //
     public function index()
     {
+        if (isset(auth()->user()->id)) {
+//           TODO middleware && construct로 바로 해결 할수 있으니 확인이 필요
+            $categories = Category::orderby('title', 'asc')->get();
 
-        $categories = Category::orderby('title', 'asc')->get();
-
-        return view('category.index')->with('categories', $categories);
+            return view('category.index')->with('categories', $categories);
+        } else {
+            return redirect('/');
+        }
     }
+
     public function view($id)
     {
-        $category = Category::find($id);
-        return view('category.view')->with('category', $category);
+        if (isset(auth()->user()->id)) {
+            $category = Category::find($id);
+            return view('category.view')->with('category', $category);
+        } else {
+            return redirect('/');
+        }
     }
+
     public function store(Request $request)
     {
-        $category = new Category;
-        $category->title =  $request->title;
-        $category->save();
+        if (isset(auth()->user()->id)) {
+            $category = new Category;
+            $category->title = $request->title;
+            $category->save();
 
-        return redirect('/category');
-    }
-    public function delete($id){
-
-        $category = Category::find($id);
-        $category->delete();
-
-        return redirect('/category');
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
     }
 
-    public function update(Request $request, $id){
+    public function delete($id)
+    {
+        if (isset(auth()->user()->id)) {
+            $category = Category::find($id);
+            $category->delete();
 
-        $category = Category::find($id);
-        $category->title =  $request->title;
-        $category->save();
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
+    }
 
-        return redirect('/category');
+    public function update(Request $request, $id)
+    {
+        if (isset(auth()->user()->id)) {
+            $category = Category::find($id);
+            $category->title = $request->title;
+            $category->save();
+
+            return redirect('/category');
+        } else {
+            return redirect('/');
+        }
     }
 }
